remove lxml-dependent code
