+++ .appveyor.yml
@@ -27,7 +27,7 @@ test_script:
   - ps: |
       $Env:VERSION = $(python mitmproxy/version.py)
       $Env:SKIP_MITMPROXY = "python -c `"print('skip mitmproxy')`""
-      tox -e wheel -- https://snapshots.mitmproxy.org/misc/lxml-3.6.0-cp35-cp35m-win32.whl
+      tox -e wheel
       tox -e rtool -- bdist
 
 deploy_script:
