+++ setup.py
@@ -70,7 +70,6 @@ setup(
         "html2text>=2016.1.8, <=2016.9.19",
         "hyperframe>=4.0.1, <5",
         "jsbeautifier>=1.6.3, <1.7",
-        "lxml>=3.5.0, <=3.6.0",  # no wheels for 3.6.1 yet.
         "Pillow>=3.2, <3.5",
         "passlib>=1.6.5, <1.8",
         "pyasn1>=0.1.9, <0.2",
